--- a/lib/graphql/backtrace.rb
+++ b/lib/graphql/backtrace.rb
@@ -2,6 +2,8 @@
 module GraphQL
   # Add GraphQL metadata to backtraces using `TracePoint`.
   #
+  # WARNING: {.enable} is not threadsafe because {GraphQL::Tracing.install} is not threadsafe.
+  #
   # @example toggling backtrace annotation
   #   # to enable:
   #   GraphQL::Backtrace.enable
@@ -28,29 +30,46 @@ module GraphQL
 
     # Implement the {GraphQL::Tracing} API.
     def trace(key, metadata)
-      pushed = false
-      case key
+      push_data = case key
       when "lex", "parse"
         # No context here, don't have a query yet
-      when "execute_multiplex"
-        clear_context_state
-        pushed = true
-        execution_context.push(metadata[:multiplex].queries)
-      when "analyze_multiplex"
-        pushed = true
-        execution_context.push(metadata[:multiplex].queries)
+        nil
+      when "execute_multiplex", "analyze_multiplex"
+        metadata[:multiplex].queries
       when "validate", "analyze_query", "execute_query", "execute_query_lazy"
-        pushed = true
-        execution_context.push(metadata[:query] || metadata[:queries])
+        metadata[:query] || metadata[:queries]
       when "execute_field", "execute_field_lazy"
-        pushed = true
-        execution_context.push(metadata[:context])
+        metadata[:context]
       else
         # Custom key, no backtrace data for this
+        nil
       end
-      yield
+
+      if push_data
+        p "Push #{key}"
+        if key == "execute_multiplex"
+          clear_context_state
+          begin
+            execution_context.push(push_data)
+            yield
+          rescue StandardError => err
+            p "Annotating #{err}"
+            annotate_error(err)
+            clear_context_state
+            # puts err.backtrace
+            raise
+          end
+        else
+          execution_context.push(push_data)
+          yield
+        end
+      else
+        yield
+      end
+
     ensure
-      if pushed
+      if push_data
+        p "Pop #{key}"
         execution_context.pop
       end
     end
@@ -90,70 +109,71 @@ module GraphQL
         backtrace_context.push(execution_context.last)
       when :return, :b_return
         backtrace_context.pop
-      when :raise
-        begin
-          err = tp.raised_exception
-          ctx = backtrace_context
-          if err.is_a?(NoMethodError)
-            # Since it's a no-method-error, the `:call`
-            # tracepoint was never entered.
-            ctx.push(execution_context.last)
-          end
+      end
+    end
 
-          ctx_offset = ctx.length - 1
-          new_backtrace = err.backtrace.each_with_index.map do |line,idx|
-            ctx_entry = idx < ctx_offset && ctx[ctx_offset - idx]
-            if ctx_entry
-              ctx_msg = format_context_message(serialize_context_entry(ctx_entry))
-              "#{line}#{ctx_msg}"
-            else
-              line
-            end
-          end
-          err.set_backtrace(new_backtrace)
-          clear_context_state
-        rescue StandardError => err
-          # We should never reach here, only if there's a bug in the block above!
-          warn("GraphQL::Backtrace bug: #{err}")
-          puts err.backtrace
+    # Format `context_entry` into a readable string.
+    # @param context_entry [Object] something that was pushed onto the stack during tracing
+    # @return [String] something to add to the backtrace entry
+    def serialize_context_entry(context_entry)
+      case context_entry
+      when GraphQL::Query::Context::FieldResolutionContext
+        ctx = context_entry
+        ctx_msg = "#{ctx.irep_node.owner_type.name}.#{ctx.field.name}"
+        if ctx.ast_node.arguments.any?
+          ctx_msg = "#{ctx_msg}(#{ctx.ast_node.arguments.map(&:to_query_string).join(", ")})"
         end
+
+        field_alias = ctx.ast_node.alias
+        if field_alias
+          ctx_msg = "#{ctx_msg} (as #{field_alias})"
+        end
+
+        ctx_msg
+      when GraphQL::Query
+        op_type = (context_entry.selected_operation && context_entry.selected_operation.operation_type) || "query"
+        op_name = context_entry.selected_operation_name || "<Anonymous>"
+        "#{op_type} #{op_name}"
+      when Array
+        context_entry.map { |i| serialize_context_entry(i) }.join(", ")
+      else
+        raise "Unexpected context entry: #{context_entry}"
+      end
+    end
+
+    # Apply styles to the string before it's added to the backtrace
+    # @param context_message [String] A prepared string for the backtrace
+    # @return [String] `context_message`, with styles added
+    def format_context_message(context_message)
+      " \e[1mGraphQL: #{context_message}\e[22m "
+    end
+
+    # Update `err`'s backtrace to include GraphQL-related annotations
+    # @param err [StandardError]
+    # @return [void]
+    def annotate_error(err)
+      ctx = backtrace_context
+      if err.is_a?(NoMethodError)
+        # Since it's a no-method-error, the `:call`
+        # tracepoint was never entered.
+        ctx.push(execution_context.last)
       end
 
-      # Format `context_entry` into a readable string.
-      # @param context_entry [Object] something that was pushed onto the stack during tracing
-      # @return [String] something to add to the backtrace entry
-      def serialize_context_entry(context_entry)
-        case context_entry
-        when GraphQL::Query::Context::FieldResolutionContext
-          ctx = context_entry
-          ctx_msg = "#{ctx.irep_node.owner_type.name}.#{ctx.field.name}"
-          if ctx.ast_node.arguments.any?
-            ctx_msg = "#{ctx_msg}(#{ctx.ast_node.arguments.map(&:to_query_string).join(", ")})"
-          end
-
-          field_alias = ctx.ast_node.alias
-          if field_alias
-            ctx_msg = "#{ctx_msg} (as #{field_alias})"
-          end
-
-          ctx_msg
-        when GraphQL::Query
-          op_type = (context_entry.selected_operation && context_entry.selected_operation.operation_type) || "query"
-          op_name = context_entry.selected_operation_name || "<Anonymous>"
-          "#{op_type} #{op_name}"
-        when Array
-          context_entry.map { |i| serialize_context_entry(i) }.join(", ")
+      ctx_offset = ctx.length - 1
+      new_backtrace = err.backtrace.each_with_index.map do |line,idx|
+        ctx_entry = idx < ctx_offset && ctx[ctx_offset - idx]
+        if ctx_entry
+          ctx_msg = format_context_message(serialize_context_entry(ctx_entry))
+          "#{line}#{ctx_msg}"
         else
-          raise "Unexpected context entry: #{context_entry}"
+          line
         end
       end
-
-      # Apply styles to the string before it's added to the backtrace
-      # @param context_message [String] A prepared string for the backtrace
-      # @return [String] `context_message`, with styles added
-      def format_context_message(context_message)
-        " \e[1mGraphQL: #{context_message}\e[22m "
-      end
+      err.set_backtrace(new_backtrace)
+    rescue StandardError => err
+      # We should never reach here, only if there's a bug in the block above!
+      warn("GraphQL::Backtrace bug: #{err}")
+      puts err.backtrace
     end
   end
 end