--- a/lib/graphql/backtrace.rb
+++ b/lib/graphql/backtrace.rb
@@ -11,20 +11,47 @@ module GraphQL
   #   GraphQL::Backtrace.disable
   #
   module Backtrace
-    GRAPHQL_GEM_PATH = File.expand_path(File.join(__FILE__, ".."))
+    # When {Backtrace} is enabled,
+    # raised errors are wrapped with `TracedError`.
+    class TracedError < GraphQL::Error
+      # @return [Array<String>] Printable backtrace of GraphQL error context
+      attr_reader :graphql_backtrace
+
+      # @return [Array<GraphQL::Execution::Multiplex, GraphQL::Query, GraphQL::Query::Context>] Objects which represent context
+      attr_reader :graphql_context
+
+      MESSAGE_TEMPLATE = <<-MESSAGE
+Unhandled error during GraphQL execution: %{cause_message}.
+Use #cause to access the original exception (including #cause.backtrace).
+
+GraphQL Backtrace:
+
+  %{graphql_backtrace}
+MESSAGE
+
+      def initialize(err, graphql_context)
+        @graphql_context = graphql_context
+        @graphql_backtrace = graphql_context.map do |key, ctx_entry|
+          "#{(key + ":").ljust(20)} #{Backtrace.serialize_context_entry(ctx_entry)}"
+        end
+        message = MESSAGE_TEMPLATE % {
+          cause_message: err.message,
+          graphql_backtrace: graphql_backtrace.map { |l| "  " + l }.join("\n"),
+        }
+        super(message)
+      end
+    end
     module_function
     # Turn on annotation
     def enable
-      clear_context_state
+      execution_context.clear
       GraphQL::Tracing.install(self)
-      TRACE_POINT.enable
       nil
     end
 
     # Turn off annotation
     def disable
       GraphQL::Tracing.uninstall(self)
-      TRACE_POINT.disable
       nil
     end
 
@@ -47,29 +74,27 @@ module GraphQL
 
       if push_data
         if key == "execute_multiplex"
-          clear_context_state
+          execution_context.clear
+          execution_context.push([key, push_data])
+          begin
+            yield
+          rescue StandardError => err
+            # This is an unhandled error from execution,
+            # Re-raise it with a GraphQL trace.
+            graphql_backtrace = execution_context.dup.uniq.reverse
+            raise TracedError.new(err, graphql_backtrace)
+          ensure
+            execution_context.clear
+          end
+        else
+          execution_context.push([key, push_data])
+          res = yield
+          execution_context.pop
+          res
         end
-        execution_context.push(push_data)
-        yield
       else
         yield
       end
-
-    ensure
-      if push_data
-        execution_context.pop
-      end
-    end
-
-    # A stack of objects, where each entry corresponds to an entry
-    # in the backtrace.
-    # The top of the stack corresponds to the top-most entry of the backtrace.
-    # The backtrace may be longer than this stack; those extra
-    # backtrace entries have no GraphQL context. (It may be longer
-    # because the backtrace is non-zero when `.enable` is called.)
-    # @return [Array<Object>]
-    def backtrace_context
-      Thread.current[:graphql_backtrace_context] ||= []
     end
 
     # A stack of objects corresponding to the GraphQL context.
@@ -79,27 +104,6 @@ module GraphQL
     # @return [Array<Object>]
     def execution_context
       Thread.current[:graphql_execution_context] ||= []
-    end
-
-    def clear_context_state
-      execution_context.clear
-      backtrace_context.clear
-    end
-
-    # - Gather info about method calls when they happen
-    # - Remove trace info for this method when it returns
-    # - Modify errors as they're being raised
-    # @api private
-    TRACE_POINT = TracePoint.new do |tp|
-      case tp.event
-      when :call, :b_call
-        backtrace_context.push(execution_context.last)
-      when :return, :b_return
-        backtrace_context.pop
-      when :raise
-        annotate_error(tp.raised_exception)
-        clear_context_state
-      end
     end
 
     # Format `context_entry` into a readable string.
@@ -129,41 +133,6 @@ module GraphQL
       else
         raise "Unexpected context entry: #{context_entry}"
       end
-    end
-
-    # Apply styles to the string before it's added to the backtrace
-    # @param context_message [String] A prepared string for the backtrace
-    # @return [String] `context_message`, with styles added
-    def format_context_message(context_message)
-      " \e[1mGraphQL: #{context_message}\e[22m "
-    end
-
-    # Update `err`'s backtrace to include GraphQL-related annotations
-    # @param err [StandardError]
-    # @return [void]
-    def annotate_error(err)
-      ctx = backtrace_context
-      if err.is_a?(NoMethodError)
-        # Since it's a no-method-error, the `:call`
-        # tracepoint was never entered.
-        ctx.push(execution_context.last)
-      end
-
-      ctx_offset = ctx.length - 1
-      new_backtrace = err.backtrace.each_with_index.map do |line,idx|
-        ctx_entry = idx < ctx_offset && ctx[ctx_offset - idx]
-        if ctx_entry
-          ctx_msg = format_context_message(serialize_context_entry(ctx_entry))
-          "#{line}#{ctx_msg}"
-        else
-          line
-        end
-      end
-      err.set_backtrace(new_backtrace)
-    rescue StandardError => err
-      # We should never reach here, only if there's a bug in the block above!
-      warn("GraphQL::Backtrace bug: #{err}")
-      puts err.backtrace
     end
   end
 end