… .travis.yml
… guides/queries/backtrace_annotations.md
… lib/graphql.rb
… lib/graphql/backtrace/table.rb
… lib/graphql/backtrace/traced_error.rb
… lib/graphql/backtrace/tracer.rb
… lib/graphql/execution/execute.rb
… lib/graphql/query/context.rb
… lib/graphql/query/variables.rb
… lib/graphql/tracing.rb
… spec/graphql/backtrace_spec.rb
… spec/graphql/query/context_spec.rb
… spec/graphql/query/variables_spec.rb
… spec/graphql/tracing/active_support_notifications_tracing_spec.rb
… spec/graphql/tracing_spec.rb
… spec/spec_helper.rb