We can use `GraphQL::Tracing` to accumulate info about execution, then re-merge it with the backtrace when errors happen. 

Also, add `ctx.backtrace` so you can print the backtrace anytime with 

```ruby
puts ctx.backtrace 
# Loc  | Field             | Object    | Arguments | Result
# 10:9 | Query.pushContext | "rootval" | {}        | 1
# 2:7  | query             | "rootval" | {}        | {backtrace: {...}, pushContext: 1}
```


## Before 

```
NoMethodError: undefined method `each' for :not_a_list:Symbol
    /Users/rmosolgo/code/graphql-ruby/lib/graphql/execution/execute.rb:192:in `resolve_value'
    /Users/rmosolgo/code/graphql-ruby/lib/graphql/execution/execute.rb:155:in `continue_resolve_field'
    /Users/rmosolgo/code/graphql-ruby/lib/graphql/execution/execute.rb:129:in `resolve_field'
    /Users/rmosolgo/code/graphql-ruby/lib/graphql/execution/execute.rb:71:in `block (2 levels) in resolve_selection'
    /Users/rmosolgo/code/graphql-ruby/lib/graphql/tracing.rb:40:in `block in trace'
    /Users/rmosolgo/code/graphql-ruby/lib/graphql/tracing.rb:71:in `call_tracers'
    /Users/rmosolgo/code/graphql-ruby/lib/graphql/tracing.rb:40:in `trace'
    /Users/rmosolgo/code/graphql-ruby/lib/graphql/execution/execute.rb:70:in `block in resolve_selection'
    /Users/rmosolgo/code/graphql-ruby/lib/graphql/execution/execute.rb:63:in `each'
    /Users/rmosolgo/code/graphql-ruby/lib/graphql/execution/execute.rb:63:in `resolve_selection'
```

## After 

```
GraphQL::Backtrace::TracedError: Unhandled error during GraphQL execution:

  undefined method `each' for :not_a_list:Symbol

Use #cause to access the original exception (including #cause.backtrace).

GraphQL Backtrace:
Loc  | Field            | Object     | Arguments | Result
1:29 | Thing.listField  | :something | {}        | (error)
1:20 | Query.field1     | nil        | {}        | {}
1:1  | query BrokenList | nil        | {}        | {}


    /Users/rmosolgo/code/graphql-ruby/lib/graphql/backtrace.rb:58:in `rescue in trace'
    /Users/rmosolgo/code/graphql-ruby/lib/graphql/backtrace.rb:61:in `trace'
    /Users/rmosolgo/code/graphql-ruby/lib/graphql/tracing.rb:73:in `call_tracers'
    /Users/rmosolgo/code/graphql-ruby/lib/graphql/tracing.rb:40:in `trace'
    /Users/rmosolgo/code/graphql-ruby/lib/graphql/execution/multiplex.rb:50:in `run_queries'
    /Users/rmosolgo/code/graphql-ruby/lib/graphql/execution/multiplex.rb:40:in `run_all'
    /Users/rmosolgo/code/graphql-ruby/lib/graphql/schema.rb:256:in `block in multiplex'
    /Users/rmosolgo/code/graphql-ruby/lib/graphql/schema.rb:560:in `with_definition_error_check'
```

## Performance impact 

There is a small perf impact, here's the output from a benchmark on my machine 

### Without backtrace 

```
~/code/graphql-ruby $ be rake bench:query
Warming up --------------------------------------
               query     2.000  i/100ms
Calculating -------------------------------------
               query     25.646  (± 3.9%) i/s -    128.000  in   5.000069s
~/code/graphql-ruby $ be rake bench:query
Warming up --------------------------------------
               query     2.000  i/100ms
Calculating -------------------------------------
               query     26.298  (± 3.8%) i/s -    132.000  in   5.023957s
~/code/graphql-ruby $ be rake bench:query
Warming up --------------------------------------
               query     2.000  i/100ms
Calculating -------------------------------------
               query     26.349  (± 3.8%) i/s -    132.000  in   5.016218s
```

### With backtrace 

```
~/code/graphql-ruby $ be rake bench:query
Warming up --------------------------------------
               query     2.000  i/100ms
Calculating -------------------------------------
               query     24.853  (± 4.0%) i/s -    124.000  in   5.007307s
~/code/graphql-ruby $ be rake bench:query
Warming up --------------------------------------
               query     2.000  i/100ms
Calculating -------------------------------------
               query     24.750  (± 8.1%) i/s -    124.000  in   5.031750s
~/code/graphql-ruby $ be rake bench:query
Warming up --------------------------------------
               query     2.000  i/100ms
Calculating -------------------------------------
               query     24.201  (± 4.1%) i/s -    122.000  in   5.058803s
```