@rmosolgo 2017-09-08
fixes #904