@eapache 2017-09-08
IIUC from looking at the usage of `TracePoint` this would intercept and annotate all exceptions within the ruby process? At least for us, this isn't suitable because GraphQL is only one piece of our server. We would much prefer to simply rescue all exceptions just before they "exit" the gem, annotate them there, and then reraise, leaving non-graphql-related exceptions untouched.