@rmosolgo 2017-09-08
Yeah, you're totally right 😬  I think I got worked up after reading the `pretty_backtrace` source! I'll refactor this into a `rescue` block which should do the job just fine. (Now, to figure out where that `rescue` block should go ...)