@rmosolgo 2017-09-12
Thanks again for the review, @eapache! I took a try at moving to `rescue`, but one problem is tracking the call stack. 

The current implementation uses a tracepoint on method & block calls & returns to maintain a stack of GraphQL context-related objects (Query, Field, etc) which parallels the call stack. Whenever the call stack changes (via call or return), this context stack is modified in the same way. 

This allows us to merge GraphQL context into the error backtrace -- starting at the top of the backtrace, there's one sister entry in the GraphQL context stack. (Eventually, the GraphQL context stack runs out, and there are several unannotated entries in the backtrace.)

The problem is that, when using `begin ... rescue`, the tracepoint registers returns _before_ the error is annotated, so the GraphQL context stack is empty (even though the error backtrace is populated). 

This is demonstrated in 1e4fe0f, which has some "logging" on the stack events. Here it is: 

```
"Push execute_multiplex"
"Push analyze_multiplex"
"Push validate"
"Pop validate"
"Push analyze_query"
"Pop analyze_query"
"Pop analyze_multiplex"
"Push execute_query"
"Push execute_field"
"Push execute_field"
"Pop execute_field"
"Pop execute_field"
"Pop execute_query"
"Annotating This is broken: Boom"
"Pop execute_multiplex"
```

Notice that _most_ of the context is popped before the error is annotated. Because of this, it's _too late_ to annotate the error! 

Contrast this with the tracepoint-based approach's output: 

```
"Push execute_multiplex"
"Push analyze_multiplex"
"Push validate"
"Pop validate"
"Push analyze_query"
"Pop analyze_query"
"Pop analyze_multiplex"
"Push execute_query"
"Push execute_field"
"Push execute_field"
"Annotating This is broken: Boom"
"Pop execute_field"
"Pop execute_field"
"Pop execute_query"
"Pop execute_multiplex"
```

Notice that `Annotating...` comes _before_ the `pop`s, so the context is still present at that point.

What can we do? I can think of two things: 

- Find another way to manage GraphQL context so that it can be re-merged into the error backtrace
- Keep using tracepoint despite the downside (and try to find a more production-ready implementation in the future) (This was done in b6d37e9a)

I think the current implementation is _good enough_ for test and development, what do you think? Or, is there another way to avoid the all-encompassing tracepoint approach?