@xuorig 2017-09-12
Small other thing to consider, although maybe not the biggest deal, is that I think some might look at the backtrace for things like exception classification. Modifying the backtrace now possibly becomes a "breaking change"?