@rmosolgo 2017-09-12
> I think some might look at the backtrace for things like exception classification

Can you help me understand what you mean by that, perhaps an example? 

Definitely don't want to break tooling!