@xuorig 2017-09-12
If we kept the path of execution in the graphql query context, could we add that on top on the stack trace at the end?

like the example you provided in the issue:

```
in field #product
in field #shop
in field #query
... rest of backtrace
```