@xuorig 2017-09-12
@rmosolgo for example we rely on backtraces to classify exception reporting into different channels / how it should be handled 🤷‍♂️