@rmosolgo 2017-09-12
> could we add that on top on the stack trace at the end?

yes, i think so.

Another option to keep them separate would be to serve `#graphql_backtrace` as a separate method. Then we coudl either add it to the backtrace or add it to the message, something like that, but leave the Ruby backtrace lines alone.