@eapache 2017-09-12
Ya, I like the idea of leaving it a separate method on the exception object and let users merge or process otherwise as they see fit.