@rmosolgo 2017-09-12
I can think of a few implementations: 

- Monkey-patch `StandardError` (:-1:)
- Rescue an error, then re-raise a `GraphQL::TracedError` (?), and make sure that `err.cause` returns the original exception 
- Stash the GraphQL context someplace else? `GraphQL::Backtrace.current`? 

Or, something else?