@eapache 2017-09-12
Wrapping and using `err.cause` makes sense; it's what that field is for.