@xuorig 2017-09-12
random ideam: what if `GraphQL::ResolutionError` (any error that happens during resolution) took `ctx` always, and had a `#graphql_backtrace` method? Then all other specific errors can inherit?