@eapache 2017-09-12
We'll still need to catch+wrap all the errors (`NameError`, `ArgumentError`) that aren't intentionally raised.