@rmosolgo 2017-09-14
Refactored to raise a wrapper error, `GraphQL::Backtrace::TracedError`. 

Currently it: 

- exposes the original error as `cause`
- exposes the GraphQL trace as `graphql_backtrace`
- includes the original error + graphql trace in `.message`
- leaves all Ruby backtraces as-is 

But I think I can do one better. Since #923, we can backtrack through `context` objects, so we could render a trace which includes arguments & returned objects, too. Let me take a crack at that.