@rmosolgo 2017-09-14
Started rendering the backtrace as a table, whadd'ya think?

```
GraphQL Backtrace:
Event             | Field                                  | Object     | Arguments           | Result
execute_field     | Thing.raiseField @ [3:13] as boomError | :something | {"message"=>"Boom"} | nil
execute_field     | Query.field1 @ [2:11]                  | nil        | {}                  | {}
execute_query     | query @ [1:9]                          | nil        | {"msg"=>"Boom"}     |
execute_multiplex | query @ [1:9]                          | nil        | {"msg"=>"Boom"}     |
```

In fact, I might depart even further from the current design. 

To render a table like this, we only need `ctx` 😎 

We can also render a table for errors from promises, since we can backtrack using ctx. I'm going to work on it a bit more.