@rmosolgo 2017-09-14
Wow, i think it can be refactored to _just_ `rescue`, using backtracking instead of managing the stack...