@rmosolgo 2017-09-15
sporadic jruby fails :/